Set the target file modification date according to Last-Modified
The combination of setting the file modification date to the value of
Last-Modified and sending If-Modified-Since when the target file already
exists enables a speed-up for the static site generation.

The local files are effectively used as a cache if the application
supports sending a 304 Not Modified in response to a request with
If-Modified-Since.