Localize all the parts of *ARGV touched by <>
The --filter option of wallflower also uses the local @ARGV idiom,
so after the first line is read the ARGV filehandle has reached EOF
(because of the use of the idiom in Wallflower::Util::_link_from_css)
and none of the following URL will be read.

The issue is that the magical <> touches all of @ARGV, $ARGV and ARGV,
so when localizing any of them, one must in fact localize *all* of them.