Add a --parallel option to wallflower, to fetch files in parallel
The child processes operate independently, and communicate via the
__SEEN__ file to figure out which URL have already been processed.

The main change is the addition of the has_seen() method, which uses
the "seen" hash in normal mode, and the __SEEN__ file in parallel mode.

However, the queue management is not great: the parent forks a child
process and hands it half of its queue. Then everyone works on their
own queue, following new links and skipping the already seen stuff.

So the work might not be distributed evenly: some child processes may
reach the end of their queue quickly (because they're exploring a deep
branch), even though there is still some work to do. They'll die and
the parent may later fork again with an updated queue.