Try to improve parallelism by sharing a list of pages to process
In addition to __SEEN__, used to share the list of already processed
pages between children, we also use a __TODO__ file to share the
to-do list.

Using the same application as before, with 1024 links and randomly
generated pages (taking one second to generate), we get the following
measurements:
with --parallel 4,  it takes about 265 seconds (256 + 9),
with --parallel 8,  it takes about 137 seconds (128 + 9),
with --parallel 16, it takes about  74 seconds (64 + 10),
with --parallel 32, it takes about  51 seconds (32 + 19),
with --parallel 64, it takes about  48 seconds (16 + 32),
with --parallel 128, it takes about 47 seconds (8 + 39).

In this case, the communication overhead is deeply worsened by the
addition of another shared and locked file.