Keep track of the number of forked children in the parent process
The previous way was more complicated (creating unique files and
counting them?), and open to a race condition (if the parent tried
to fork before the child had created the file).

Keeping a counter in the parent process is much simpler, faster
and avoids this kind of race condition entirely.

We still need the files (which are deleted when the child processes
terminate), to make sure all children have effectively terminated.

 * * *

Here are some basic speed measurements, done on a 4 cpu laptop, using the
following basic application, where each page takes one second to render,
and the entire site is made of 1024 pages:

    my $max = shift || 1024;
    my $app = sub {
        my ($env) = @_;
        sleep 1;
        $env->{REQUEST_URI} =~ m{/(\d+)};
        srand( $1 || 0 );
        my $html = join "\n",
          '<html><head></head><body>',
          map ( qq{<a href="/$_">$_</a>}, map int rand $max, 0 .. rand $max ),
          "</body><html>";
        [
            200,
            [
                'Content-Type'   => 'text/html',
                'Content-Length' => length($html)
            ],
            [$html]
        ];
    };

Running wallflower in the regular serial mode takes 1049 seconds (1024 + 25),
with --parallel 4, it takes about 265 seconds (256 + 9),
with --parallel 8, it takes about 133 seconds (128 + 5),
with --parallel 16, it takes about 68 seconds (64 + 4),
with --parallel 32, it takes about 37 seconds (32 + 5),
with --parallel 64, it takes about 22 seconds (16 + 6),
with --parallel 128, it takes about 19 seconds (8 + 11).

With too many child processes, it seems that the communication overhead
becomes larger.