Improve parallelism when sharing a __TODO__ file
To avoid the latency caused by all the children waiting on and updating
a shared __TODO__ file, in this version only the parent process writes
the main __TODO__ file, while all children write their to-do list in
temporary files.

When a process empties its to-do list, it will read the main __TODO__
to obtain new links to explore. If that doesn't yield anything not
already seen, it will stop.

The main __TODO__ is written by aggregating all the children to-do lists
with the parent's one, and filtering it using the list of seen links.
No locking is needed when updating the file, because it is simply
replaced by the new version (a child reading the previous version will
keep reading it).

The parent forks new kids as early as possible, as long as its to-do
list is not empty. Each child process starts with an item from the
parent's to-do.

Using the same application as before, with 1024 links and randomly
generated pages (taking one second to generate), we get the following
measurements:
with --parallel 4,  it takes about 265 seconds (256 + 9),
with --parallel 8,  it takes about 133 seconds (128 + 5),
with --parallel 16, it takes about 67 seconds (64 + 3),
with --parallel 32, it takes about 35 seconds (32 + 3),
with --parallel 64, it takes about 19 seconds (16 + 3),
with --parallel 128, it takes about 14 seconds (8 + 6),
with --parallel 256, it takes about 13 seconds (4 + 9).

This is better than the previous two versions, and significantly better
than the first version with a shared __TODO__. Still, when there are
too many processes competing to read and update the __SEEN__ file,
the communication overhead nullifies the gain from parallelism.