Keep the content of the old __TODO__ when rebuilding it
When combining --parallel and --filter, a race condition between
the parent and the children to access __TODO__ would let the
parent wipe out the content of __TODO__ if it hadn't been picked
up by one of the children before.

The fix is easy: also use the existing content of __TODO__ when
rebuilding it.